@@ -9,24 +9,24 @@
   "dependencies": {
     "@react-spring/core": "^9.2.4",
     "@react-spring/three": "^9.2.4",
-    "@react-three/drei": "^7.2.1",
-    "lodash-es": "^4.17.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-merge-refs": "^1.1.0",
-    "react-router-dom": "^5.2.0",
+    "@react-three/drei": "^7.6.1",
     "@use-gesture/react": "latest",
-    "styled-components": "^5.3.0",
-    "three-stdlib": "^2.3.1",
+    "lodash-es": "^4.17.21",
+    "react": "^18.0.0-alpha-1314299c7-20210901",
+    "react-dom": "^18.0.0-alpha-1314299c7-20210901",
+    "react-merge-refs": "^1.1.0",
+    "react-router-dom": "^5.2.1",
+    "styled-components": "^5.3.1",
+    "three-stdlib": "^2.4.0",
     "use-error-boundary": "^2.0.6",
-    "zustand": "^3.5.7"
+    "zustand": "^3.5.10"
   },
   "devDependencies": {
-    "@types/react": "^17.0.14",
+    "@types/react": "^17.0.19",
     "@types/react-router-dom": "^5.1.8",
-    "@types/styled-components": "^5.1.11",
-    "@vitejs/plugin-react-refresh": "^1.3.5",
-    "typescript": "^4.3.5",
-    "vite": "^2.4.3"
+    "@types/styled-components": "^5.1.13",
+    "@vitejs/plugin-react-refresh": "^1.3.6",
+    "typescript": "^4.4.2",
+    "vite": "^2.5.3"
   }
 }
