@@ -36,45 +36,45 @@
     "vers": "yarn changeset version"
   },
   "devDependencies": {
-    "@babel/core": "7.14.8",
-    "@babel/preset-env": "7.14.8",
+    "@babel/core": "7.15.4",
+    "@babel/preset-env": "7.15.4",
     "@babel/preset-react": "7.14.5",
-    "@babel/preset-typescript": "^7.14.5",
-    "@changesets/changelog-git": "^0.1.7",
-    "@changesets/cli": "^2.16.0",
-    "@preconstruct/cli": "^2.1.0",
+    "@babel/preset-typescript": "^7.15.0",
+    "@changesets/changelog-git": "^0.1.8",
+    "@changesets/cli": "^2.17.0",
+    "@preconstruct/cli": "^2.1.1",
     "@testing-library/react": "^12.0.0",
-    "@types/jest": "^26.0.24",
+    "@types/jest": "^27.0.1",
     "@types/lodash-es": "^4.17.4",
-    "@types/react": "^17.0.14",
+    "@types/react": "^17.0.19",
     "@types/react-dom": "^17.0.9",
-    "@types/react-reconciler": "^0.26.2",
+    "@types/react-reconciler": "^0.26.3",
     "@types/react-test-renderer": "^17.0.1",
     "@types/scheduler": "^0.16.2",
-    "@types/three": "^0.130.0",
-    "@typescript-eslint/eslint-plugin": "^4.28.4",
-    "@typescript-eslint/parser": "^4.28.4",
-    "eslint": "^7.31.0",
+    "@types/three": "^0.131.0",
+    "@typescript-eslint/eslint-plugin": "^4.30.0",
+    "@typescript-eslint/parser": "^4.30.0",
+    "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-import-resolver-alias": "^1.1.2",
-    "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-jest": "^24.3.6",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react": "^7.24.0",
+    "eslint-plugin-import": "^2.24.2",
+    "eslint-plugin-jest": "^24.4.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.25.1",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "husky": "^7.0.1",
-    "jest": "^27.0.6",
-    "jest-cli": "^27.0.6",
-    "lint-staged": "^11.0.1",
+    "husky": "^7.0.2",
+    "jest": "^27.1.0",
+    "jest-cli": "^27.1.0",
+    "lint-staged": "^11.1.2",
     "prettier": "^2.3.2",
     "pretty-quick": "^3.1.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "regenerator-runtime": "^0.13.7",
-    "three": "^0.130.1",
-    "three-stdlib": "^2.3.1",
-    "ts-jest": "^27.0.3",
-    "typescript": "^4.3.5"
+    "react": "^18.0.0-alpha-1314299c7-20210901",
+    "react-dom": "^18.0.0-alpha-1314299c7-20210901",
+    "regenerator-runtime": "^0.13.9",
+    "three": "^0.132.2",
+    "three-stdlib": "^2.4.0",
+    "ts-jest": "^27.0.5",
+    "typescript": "^4.4.2"
   },
   "dependencies": {}
 }
