@@ -40,21 +40,20 @@
     "prebuild": "cp ../../readme.md readme.md"
   },
   "dependencies": {
-    "@babel/runtime": "^7.13.10",
+    "@babel/runtime": "^7.15.4",
     "react-merge-refs": "^1.1.0",
-    "react-reconciler": "^0.26.2",
-    "react-three-fiber": "0.0.0-deprecated",
+    "react-reconciler": "^0.27.0-alpha-1314299c7-20210901",
     "react-use-measure": "^2.0.4",
     "resize-observer-polyfill": "^1.5.1",
-    "scheduler": "^0.20.2",
+    "scheduler": "^0.21.0-alpha-1314299c7-20210901",
     "use-asset": "^1.0.4",
     "utility-types": "^3.10.0",
-    "zustand": "^3.5.1"
+    "zustand": "^3.5.10"
   },
   "peerDependencies": {
-    "react": ">=17.0",
-    "react-dom": ">=17.0",
-    "three": ">=0.126"
+    "react": ">=18.0",
+    "react-dom": ">=18.0",
+    "three": ">=0.132"
   },
   "peerDependenciesMeta": {
     "react-dom": {
